requirements.txt
backcall==0.1.0
certifi==2019.9.11
chardet==3.0.4
confusable-homoglyphs==3.2.0
decorator==4.4.1
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.7
django-allauth==0.40.0
django-crispy-forms==1.8.0
django-extensions==2.2.5
django-registration==3.0.1
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django-webpack-loader==0.6.0
django-widget-tweaks==1.4.5
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
gunicorn==20.0.0
idna==2.8
ipython==7.9.0
ipython-genutils==0.2.0
jedi==0.15.1
numpy==1.17.4
oauthlib==3.1.0
pandas==0.25.3
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==2.0.10
psycopg2==2.8.4
ptyprocess==0.6.0
Pygments==2.4.2
PyJWT==1.7.1
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
six==1.13.0
sqlparse==0.3.0
traitlets==4.3.3
urllib3==1.25.7
wcwidth==0.1.7
whitenoise==4.1.4
